Changes before test PyPI
Ensuring pytest works
Renaming new bidirectional algorithm to NBAstar
Switching over to pyproject instead of setup.py
